Fix bench definitions
The problem was that names of the suite are used for generating certain
files. Since the name included a `/`, the generated files were being
created as directories instead of flat files. I think a fix for this is
to make sure we escape `/` in the names before we generate files. But
the short term fix here is just to ensure that the names of the bench
suites don't look like nested file paths.