Don't pretty print XML
We generate XML files to configure benchexec.  We had been
pretty-printing these files, to make it easier to read the generated
files for debugging etc. However, benchexec doesn't play well with
pretty printed xml (see
https://github.com/informalsystems/apalache-bench/pull/54#issuecomment-1166634666).
So, as a workaround, this PR drops the pretty printing and just dumps
the XML directly.